Refactor email config.
